Reduced the size of the Squad_Strength class so that the AI elements need less memory: ctp2_code/ai/ctpai.cpp ctp2_code/ai/mapanalysis/mapanalysis.cpp ctp2_code/ai/strategy/agents/agent.cpp ctp2_code/ai/strategy/goals/Goal.cpp ctp2_code/ai/strategy/scheduler/scheduler.cpp ctp2_code/ai/strategy/squads/squad_Strength.cpp ctp2_code/ai/strategy/squads/squad_Strength.h ctp2_code/gs/world/cellunitlist.cpp ctp2_code/gs/world/cellunitlist.h
Fixed 64 bit portability problems:
ctp2_code/robot/pathing/AVLHeap.cpp
ctp2_code/ui/interface/GovernmentTab.cpp


git-svn-id: http://ctp2.darkdust.net/anonsvn/trunk@1071 34d2fc0c-33df-0310-8a34-85f8040ba873

+++ ctp2_code/ai/strategy/agents/agent.cpp
@@ -226,13 +226,13 @@ const Squad_Strength & Agent::Compute_Squad_Strength()
 
 	sint16 defense_count;
 	sint16 ranged_count;
-	double attack_strength;
-	double defense_strength;
-	double ranged_strength;
-	double bombard_land_strength;
-	double bombard_sea_strength;
-	double bombard_air_strength;
-	double total_value;
+	float  attack_strength;
+	float  defense_strength;
+	float  ranged_strength;
+	float  bombard_land_strength;
+	float  bombard_sea_strength;
+	float  bombard_air_strength;
+	float  total_value;
 	m_army->ComputeStrength(attack_strength,
 	                        defense_strength,
 	                        ranged_strength,
