+++ ctp2_code/ai/strategy/scheduler/scheduler.cpp
@@ -1936,8 +1936,8 @@ void Scheduler::Assign_Garrison()
 
 			sint16 defense_count;
 			sint16 tmp_count;
-			double tmp;
-			double defense_strength;
+			float  tmp;
+			float  defense_strength;
 			army->ComputeStrength(tmp,
 			                      defense_strength,
 			                      tmp,
